refactor(nft-media): use custom hook to get media type
